requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.3
blinker==1.7.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
conda @ file:///croot/conda_1696257509808/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1696257536937/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
configobj==5.0.8
configparser==6.0.0
contourpy @ file:///work/ci_py311/contourpy_1676827066340/work
cryptography @ file:///croot/cryptography_1694444244250/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.12.2
distlib==0.3.7
etelemetry==0.3.1
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.25.0
frozenlist==1.4.0
future==0.18.3
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna @ file:///work/ci_py311/idna_1676822698822/work
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib @ file:///croot/joblib_1685113087166/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver @ file:///work/ci_py311/kiwisolver_1676827230232/work
langdetect==1.0.9
libmambapy @ file:///croot/mamba-split_1694187754698/work/libmambapy
looseversion==1.3.0
lru-dict==1.2.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib @ file:///croot/matplotlib-suite_1698692105134/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mollie-api-python==3.6.0
multidict==6.0.4
munkres==1.1.4
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp311-cp311-linux_x86_64.whl#sha256=184178a893ff0b5e0cb1b0b7a0dddff49dc91424c4e4a01e29bcb05f27a698a6
oauthlib==3.2.2
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.3
parsimonious==0.9.0
pathlib==1.0.1
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.11.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
ply==3.11
priority==2.0.0
protobuf==4.25.1
prov==2.0.0
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pydot==1.4.2
PyJWT==2.8.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///work/ci_py311/pyparsing_1677811559502/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.1.0
pyxnat==1.6
quart==0.18.3
quart-cors==0.7.0
quart-sqlalchemy==2.0.0
rdflib==7.0.0
referencing==0.31.0
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.13.1
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
scikit-learn @ file:///croot/scikit-learn_1684954695550/work
scipy @ file:///croot/scipy_1696543286448/work/dist/scipy-1.11.3-cp311-cp311-linux_x86_64.whl#sha256=f14db5b9e7e3d9de204ef7756f1a1f40f0983293019ed8eacb32191e2f658f9f
simplejson==3.19.2
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toolz==0.12.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm @ file:///croot/tqdm_1679561862951/work
traits==6.3.2
truststore @ file:///croot/truststore_1695244293384/work
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
uniswap-python==0.7.1
urllib3==2.0.7
uvicorn==0.24.0.post1
virtualenv==20.24.6
web3==6.11.3
websockets==12.0
Werkzeug==2.3.0
wsproto==1.2.0
yarl==1.9.2
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
